Getting Green tech articles
In this code BeautifulSoup is used to get articles about green tech from greentechmedia.com. These articles are news articles and not research articles. The resulting data frame consists of the article title, article date, article author, article content. There are 364 articles in this dataframe, each stored as an observation.